cleaning and refactorings
